mlops-service/requirements.txt
python-dotenv>=1.0.0
PyYAML>=6.0
loguru>=0.7.0
python-box>=7.1.1
pydantic>=2.7.4
pydantic-settings>=2.2.0

pandas>=2.0
numpy>=1.25
scikit-learn>=1.3
seaborn>=0.12
matplotlib>=3.7
xgboost>=2.0.0

Pillow>=10.0
opencv-python-headless>=4.8
albumentations>=1.3

torch>=2.2.0
torchvision>=0.17.0
timm>=0.9
datasets>=2.18.0
transformers>=4.40.0

dagshub>=0.3.25
mlflow>=2.7
dvc[http]>=3.49.0

evidently>=0.4
prometheus-client>=0.20.0

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0

pytest>=8.0
pytest-asyncio>=0.23
pytesseract>=0.3.10

python-multipart>=0.0.6

